Update Donkey Kong Revamp.py
This simple error fix from int to str fixed the issue where the wrong button would be displayed and you could buy luigi again, even if the text file said "True" - i.e. already bought.

This is because the true/ false from the text file was not being saved properly as it was trying to save it as an integer when it is a string