Update Donkey Kong Revamp.py
Figured out how inheritance actually works and now the game is essentially finished with everything working exactly as intended. This fix allowed my to complete the appliance of the luigi skin onto the player by using a child class of the original player class. This new super function ensures that all of the init function is inherited from the player class, and then we simply "edit" the contents of the self.image attribute instead of writing the entire init function again - which didn't work anyway