Update Donkey Kong Revamp.py
All works well except for a strange error/ bug where the luigi button seems to "skip" straight to the applyskin() method even before the yes confirmation button is pressed - which should be impossible. Also having difficulty with the smoothness of the button and exiting back to the shop once the button has been pressed for some reason.